+++ Projeto/app/src/main/java/com/gui_rei/tempopreparar/rest/days/data/Temperature.java
@@ -8,10 +8,10 @@ public class Temperature {
 
     @SerializedName("min")
     @Expose
-    private Integer min;
+    private Float min;
     @SerializedName("max")
     @Expose
-    private Integer max;
+    private Float max;
     @SerializedName("morning")
     @Expose
     private Morning morning;
@@ -22,19 +22,19 @@ public class Temperature {
     @Expose
     private Night night;
 
-    public Integer getMin() {
+    public Float getMin() {
         return min;
     }
 
-    public void setMin(Integer min) {
+    public void setMin(Float min) {
         this.min = min;
     }
 
-    public Integer getMax() {
+    public Float getMax() {
         return max;
     }
 
-    public void setMax(Integer max) {
+    public void setMax(Float max) {
         this.max = max;
     }
